// sanity.types.ts
export type Slide = {
  _type: "slide";
  text?: string;
};

export type Carousel = {
  _id: string;
  _type: "carousel";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  title?: string;
  backgroundColor?: Color;
  autoPlay?: boolean;
  infiniteLoop?: boolean;
  showArrows?: boolean;
  swipeable?: boolean;
  interval?: number;
  transitionTime?: number;
  slides?: Array<{
    _key: string;
  } & Slide>;
  navigationItems?: Array<{
    label?: string;
    link?: string;
    _type: "navItem";
    _key: string;
  }>;
  exclusiveTag?: string;
  headerColor?: Color;
  heroSlides?: Array<{
    image?: {
      asset?: {
        _ref: string;
        _type: "reference";
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
      };
      hotspot?: SanityImageHotspot;
      crop?: SanityImageCrop;
      _type: "image";
    };
    text?: string;
    _key: string;
  }>;
};

export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};
// ...
export type Color = {
  _type: "color";
  hex?: string;
  alpha?: number;
  hsl?: HslaColor;
  hsv?: HsvaColor;
  rgb?: RgbaColor;
};

export type RgbaColor = {
  _type: "rgbaColor";
  r?: number;
  g?: number;
  b?: number;
  a?: number;
};

export type HsvaColor = {
  _type: "hsvaColor";
  h?: number;
  s?: number;
  v?: number;
  a?: number;
};

export type HslaColor = {
  _type: "hslaColor";
  h?: number;
  s?: number;
  l?: number;
  a?: number;
};

export type AllSanitySchemaTypes = SanityImagePaletteSwatch | SanityImagePalette | SanityImageDimensions | SanityFileAsset | Geopoint | Slug | Slide | Carousel | SanityImageCrop | SanityImageHotspot | SanityImageAsset | SanityAssetSourceData | SanityImageMetadata | Color | RgbaColor | HsvaColor | HslaColor;
export declare const internalGroqTypeReferenceTo: unique symbol;